Split core functionality and main function so that setup.py can import __version__
